Reorganize Driver3 translators
Reorganize Driver3 translators - classes that take an object returned
by Java Driver 3.x and return a scylla-cdc-java library representation
(and the reverse way). Make those classes private, because if we shade
them, they will become useless to use with non-shaded Java Driver 
objects.

A consequence of making those translators private is that the replicator
must have its own copy of this class.
