+++ scylla-cdc-replicator/src/main/java/com/scylladb/cdc/replicator/operations/delete/PartitionDeleteOperationHandler.java
@@ -6,8 +6,8 @@ import com.datastax.driver.core.Session;
 import com.datastax.driver.core.TableMetadata;
 import com.datastax.driver.core.querybuilder.Delete;
 import com.datastax.driver.core.querybuilder.QueryBuilder;
-import com.scylladb.cdc.cql.driver3.Driver3FromLibraryTranslator;
 import com.scylladb.cdc.model.worker.RawChange;
+import com.scylladb.cdc.replicator.driver3.Driver3FromLibraryTranslator;
 import com.scylladb.cdc.replicator.operations.ExecutingPreparedStatementHandler;
 
 import static com.datastax.driver.core.querybuilder.QueryBuilder.*;
