+++ scylla-cdc-replicator/src/main/java/com/scylladb/cdc/replicator/operations/preimage/PreImageOperationHandler.java
@@ -21,12 +21,12 @@ import com.datastax.driver.core.TableMetadata;
 import com.datastax.driver.core.TypeCodec;
 import com.datastax.driver.core.querybuilder.QueryBuilder;
 import com.datastax.driver.core.querybuilder.Select;
-import com.scylladb.cdc.cql.driver3.Driver3FromLibraryTranslator;
-import com.scylladb.cdc.cql.driver3.Driver3ToLibraryTranslator;
 import com.scylladb.cdc.model.worker.ChangeSchema;
 import com.scylladb.cdc.model.worker.RawChange;
 import com.scylladb.cdc.model.worker.cql.Cell;
 import com.scylladb.cdc.model.worker.cql.Field;
+import com.scylladb.cdc.replicator.driver3.Driver3FromLibraryTranslator;
+import com.scylladb.cdc.replicator.driver3.Driver3ToLibraryTranslator;
 import com.scylladb.cdc.replicator.operations.CdcOperationHandler;
 import com.scylladb.cdc.replicator.operations.FutureUtils;
