+++ scylla-cdc-replicator/src/main/java/com/scylladb/cdc/replicator/operations/update/PreparedUpdateOperationHandler.java
@@ -7,8 +7,8 @@ import com.datastax.driver.core.Session;
 import com.datastax.driver.core.TableMetadata;
 import com.datastax.driver.core.querybuilder.QueryBuilder;
 import com.datastax.driver.core.querybuilder.Update;
-import com.scylladb.cdc.cql.driver3.Driver3FromLibraryTranslator;
 import com.scylladb.cdc.model.worker.RawChange;
+import com.scylladb.cdc.replicator.driver3.Driver3FromLibraryTranslator;
 import com.scylladb.cdc.replicator.operations.ExecutingPreparedStatementHandler;
 
 import java.util.HashSet;
